refactor: 패키지 구조 변경
- 도메인, 저장소 계층에서 기존보다 더 명확한 구조를 가집니다.
