+++ src/main/java/roomescape/business/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.business;
+package roomescape.business.domain;
 
 import java.time.LocalDate;
