+++ src/main/java/roomescape/business/domain/ReservationTheme.java
@@ -1,4 +1,4 @@
-package roomescape.business;
+package roomescape.business.domain;
 
 public final class ReservationTheme {
