+++ src/main/java/roomescape/business/service/ReservationService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.business.Reservation;
-import roomescape.business.ReservationTheme;
-import roomescape.business.ReservationTime;
+import roomescape.business.domain.Reservation;
+import roomescape.business.domain.ReservationTheme;
+import roomescape.business.domain.ReservationTime;
 import roomescape.exception.ReservationException;
 import roomescape.exception.ReservationThemeException;
 import roomescape.exception.ReservationTimeException;
