+++ src/main/java/roomescape/persistence/jdbc/JdbcMemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence;
+package roomescape.persistence.jdbc;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -7,7 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.business.Member;
+import roomescape.business.domain.Member;
+import roomescape.persistence.MemberRepository;
 
 @Repository
 public class JdbcMemberRepository implements MemberRepository {
