+++ src/test/java/roomescape/business/ReservationTimeTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.business.domain.ReservationTime;
 
 class ReservationTimeTest {
