+++ src/test/java/roomescape/business/service/ReservationTimeServiceTest.java
@@ -10,9 +10,9 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.business.Reservation;
-import roomescape.business.ReservationTheme;
-import roomescape.business.ReservationTime;
+import roomescape.business.domain.Reservation;
+import roomescape.business.domain.ReservationTheme;
+import roomescape.business.domain.ReservationTime;
 import roomescape.exception.ReservationTimeException;
 import roomescape.persistence.ReservationRepository;
 import roomescape.persistence.ReservationTimeRepository;
