+++ src/test/java/roomescape/persistence/fakerepository/FakeReservationRepository.java
@@ -6,7 +6,7 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.business.Reservation;
+import roomescape.business.domain.Reservation;
 import roomescape.persistence.ReservationRepository;
 
 @Repository
