+++ src/test/java/roomescape/presentation/ReservationControllerTest.java
@@ -14,9 +14,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import roomescape.business.Reservation;
-import roomescape.business.ReservationTheme;
-import roomescape.business.ReservationTime;
+import roomescape.business.domain.Reservation;
+import roomescape.business.domain.ReservationTheme;
+import roomescape.business.domain.ReservationTime;
 import roomescape.exception.ErrorResponseDto;
 import roomescape.persistence.fakerepository.FakeReservationRepository;
 import roomescape.persistence.fakerepository.FakeReservationThemeRepository;
