+++ src/test/java/roomescape/presentation/ReservationThemeControllerTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.server.LocalServerPort;
-import roomescape.business.ReservationTheme;
+import roomescape.business.domain.ReservationTheme;
 import roomescape.exception.ErrorResponseDto;
 import roomescape.persistence.fakerepository.FakeReservationThemeRepository;
 import roomescape.presentation.dto.ReservationThemeResponseDto;
